requirements.txt
google-cloud-storage==2.14.0
google-cloud-bigquery==3.14.1
pandas==2.1.4
python-dotenv==1.0.0
pyarrow==14.0.1 